pyproject.toml
[build-system]
requires = ["setuptools>=82.0.1"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-rpg-engine"
version = "0.1.0"
description = "A Python project using pygame and injector"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
dependencies = [
    "injector>=0.22.0",
    "pygame==2.6.1",
    "pytmx",
    "PyYAML==6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
testpaths = ["tests", "engine"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v -x"

[tool.setuptools.packages.find]
where = ["."]
include = ["engine"]

[tool.setuptools.package-data]
"engine" = ["../rusted_kingdoms/*", "../rusted_kingdoms/**/*"]
